rewrite code on typescript

+++ src/js/index.js
@@ -1,1 +1,5 @@
-function requireAll(r){r.keys().forEach(r);}requireAll(require.context('../assets/img/icons/',true,/\.svg$/));$('.js-range-slider-4').myPlugin({type:'double',flag:true});
+function requireAll(r) {
+	r.keys().forEach(r);
+}
+requireAll(require.context('../assets/img/icons/', true, /\.svg$/));
+$('.js-range-slider-4').myPlugin({ type: 'single', flag: true });
